internal/usecase/create_client/create_client.go
package createclient

import (
	"fmt"
	"time"

	"github.com/nironwp/ms-wallet/internal/entity"
	"github.com/nironwp/ms-wallet/internal/gateway"
)

type CreateClientInputDTO struct {
	Name  string
	Email string
}

type CreateClientOutputDTO struct {
	ID        string
	Name      string
	Email     string
	CreatedAt time.Time
	UpdatedAt time.Time
}

type CreateClientUseCase struct {
	ClientGateway gateway.ClientGateway
}

func NewCreateClientUseCase(clientGateway gateway.ClientGateway) *CreateClientUseCase {
	return &CreateClientUseCase{
		ClientGateway: clientGateway,
	}
}

func (u *CreateClientUseCase) Execute(input CreateClientInputDTO) (*CreateClientOutputDTO, error) {
	client, err := entity.NewClient(input.Name, input.Email)
	fmt.Println(client)
	fmt.Println("Error ocurred", err)
	if err != nil {
		fmt.Println("Error while creating client")
		return nil, err
	}

	err = u.ClientGateway.Save(client)
	if err != nil {
		fmt.Println("Error while saving", err)
		return nil, err
	}

	return &CreateClientOutputDTO{
		ID:        client.ID,
		Name:      client.Name,
		Email:     client.Email,
		CreatedAt: client.CreatedAt,
		UpdatedAt: client.UpdatedAt,
	}, nil
}
